+++ addons/main/ballistics_compat/script_component.hpp
@@ -1,0 +1,2 @@
+#define SUBCOMPONENT ballistics_compat
+#include "..\script_component.hpp"