+++ addons/main/flashsuppressors_compat/script_component.hpp
@@ -1,0 +1,2 @@
+#define SUBCOMPONENT flashsuppressors_compat
+#include "..\script_component.hpp"