+++ addons/main/logistics_uavbattery_compat/config.cpp
@@ -1,13 +1,13 @@
 #include "script_component.hpp"
 
 class CfgPatches {
-    class ADDON {
-        name = COMPONENT_NAME;
+    class SUBADDON {
+        addonRootClass = QUOTE(ADDON);
         units[] = {};
         weapons[] = {};
         requiredVersion = REQUIRED_VERSION;
         requiredAddons[] = {
-            "ace_main"
+            "ace_logistics_uavbattery"
         };
         author = "johnb43";
         authors[] = {
@@ -16,5 +16,15 @@ class CfgPatches {
         url = "https://github.com/johnb432/Weapons-Balance-ACE";
         skipWhenMissingDependencies = 1;
         VERSION_CONFIG;
+    };
+};
+
+class CfgWeapons {
+    class CBA_MiscItem_ItemInfo;
+    class ACE_ItemCore;
+    class ACE_UAVBattery: ACE_ItemCore {
+        class ItemInfo: CBA_MiscItem_ItemInfo {
+            mass = 8;
+        };
     };
 };