+++ addons/main/logistics_uavbattery_compat/script_component.hpp
@@ -1,0 +1,2 @@
+#define SUBCOMPONENT logistics_uavbattery_compat
+#include "..\script_component.hpp"