+++ addons/main/medical_treatment_compat/script_component.hpp
@@ -1,0 +1,2 @@
+#define SUBCOMPONENT medical_treatment_compat
+#include "..\script_component.hpp"