+++ addons/main/nightvision_compat/script_component.hpp
@@ -1,0 +1,2 @@
+#define SUBCOMPONENT nightvision_compat
+#include "..\script_component.hpp"