+++ addons/main/realisticnames_compat/script_component.hpp
@@ -1,0 +1,2 @@
+#define SUBCOMPONENT realisticnames_compat
+#include "..\script_component.hpp"